refactor(quantum_gate): use dispatching for get_inverse and get_operator
The Gate struct was changed to an abstract type. Every type of gate now has its own struct which is a child of the Gate abstract type.
The operator is no longer stored inside the struct. This speeds up gate contruction.
The use of dispatching for get_inverse improves the extensibility of the function.

Issue #39